refactor(search): extract _FuelBadge from all_prices_station_card
Move the private _FuelBadge widget into a sibling part file
(all_prices_station_card_parts.dart) so the main file stays under the
300-LOC budget. Behavior is unchanged: the widget remains library-private
via Dart's part/part-of mechanism, the public AllPricesStationCard API is
untouched, and all existing widget tests pass without modification.

Refs #563 phase: all_prices_station_card
